Rename Python libraries from seattleflu.* to id3c.*
This started as disparate work specific to the Seattle Flu Study without
a proper name.  There's always been a goal of using this software for
Nextstrain and other projects/studies, and we're in the first stages of
drawing lines between core ID3C and project-specific extensions.
