+++ bin/id3c
@@ -4,19 +4,19 @@ set -euo pipefail
 # Find our project root.
 root="$(cd "$(dirname "$0")/.." && pwd)"
 
-# Add our lib dir to the module search path so we can load seattleflu.db.
+# Add our lib dir to the module search path so we can load id3c.db.
 locallib="$root/lib"
 test -d "$locallib"
 export PYTHONPATH="$locallib"
 
 if [[ -n ${PIPENV_ACTIVE:-} ]]; then
     # We are already running under `pipenv run` or `pipenv shell`.
-    exec python3 -m seattleflu.db.cli "$@"
+    exec python3 -m id3c.db.cli "$@"
 else
     # We need to run under pipenv, so find our Pipfile.
     pipfile="$root/Pipfile"
     test -e "$pipfile"
     export PIPENV_PIPFILE="$pipfile"
 
-    exec pipenv run python3 -m seattleflu.db.cli "$@"
+    exec pipenv run python3 -m id3c.db.cli "$@"
 fi
