+++ lib/id3c/api/static/index.html
