+++ lib/id3c/db/cli/__main__.py
@@ -1,0 +1,7 @@
+"""
+Invoked via ``python -m id3c.db.cli``.
+"""
+
+if __name__ == "__main__":
+    from id3c.db.cli import cli
+    cli(prog_name = "id3c")
