+++ lib/id3c/db/cli/command/etl/clinical.py
@@ -6,9 +6,9 @@ import logging
 import re
 from datetime import datetime, timezone
 from typing import Any, Mapping, Optional
-from seattleflu.db import find_identifier
-from seattleflu.db.session import DatabaseSession
-from seattleflu.db.datatypes import Json
+from id3c.db import find_identifier
+from id3c.db.session import DatabaseSession
+from id3c.db.datatypes import Json
 from . import (
     etl,
