+++ lib/id3c/db/cli/command/etl/enrollments.py
@@ -7,9 +7,9 @@ from datetime import datetime, timezone
 from itertools import groupby
 from operator import itemgetter
 from typing import Any, Optional
-from seattleflu.db import find_identifier
-from seattleflu.db.session import DatabaseSession
-from seattleflu.db.datatypes import Json
+from id3c.db import find_identifier
+from id3c.db.session import DatabaseSession
+from id3c.db.datatypes import Json
 from . import etl, find_or_create_site, find_location, upsert_individual, upsert_encounter, upsert_location, upsert_encounter_location
