+++ lib/id3c/db/cli/command/longitudinal.py
@@ -11,11 +11,11 @@ import click
 import logging
 import numpy as np
 import pandas as pd
-import seattleflu.db as db
+import id3c.db as db
 from typing import List
 from itertools import combinations
-from seattleflu.db.session import DatabaseSession
-from seattleflu.db.cli import cli
+from id3c.db.session import DatabaseSession
+from id3c.db.cli import cli
 from . import (
     add_metadata,
     barcode_quality_control,
