+++ lib/id3c/db/cli/command/sequence_read_set.py
@@ -12,9 +12,9 @@ from collections import defaultdict
 from pathlib import Path
 from typing import Dict, Optional
 from urllib.parse import urljoin
-from seattleflu.db.cli import cli
-from seattleflu.db.session import DatabaseSession
-from seattleflu.db.datatypes import as_json
+from id3c.db.cli import cli
+from id3c.db.session import DatabaseSession
+from id3c.db.datatypes import as_json
 
 
 LOG = logging.getLogger(__name__)
