+++ lib/id3c/db/cli/command/user.py
@@ -3,11 +3,11 @@ List and create users.
 """
 import click
 import logging
-import seattleflu.db as db
+import id3c.db as db
 from operator import attrgetter
 from typing import List
-from seattleflu.db.session import DatabaseSession
-from seattleflu.db.cli import cli
+from id3c.db.session import DatabaseSession
+from id3c.db.cli import cli
 
 
 LOG = logging.getLogger(__name__)
