+++ lib/id3c/db/types.py
