+++ lib/id3c/labelmaker.py
