--- lib/seattleflu/db/cli/__main__.py
@@ -1,7 +1,0 @@
-"""
-Invoked via ``python -m seattleflu.db.cli``.
-"""
-
-if __name__ == "__main__":
-    from seattleflu.db.cli import cli
-    cli(prog_name = "id3c")
