+++ wsgi.py
@@ -2,11 +2,11 @@ from sys import path
 from pathlib import Path
 
 # Add our containing repo's lib directory to the Python module search path so
-# that we can load seattleflu.api.
+# that we can load id3c.api.
 libpath = Path(__file__).parent / "lib"
 assert libpath.is_dir()
 
 path.insert(0, str(libpath))
 
-from seattleflu.api import create_app
+from id3c.api import create_app
 application = create_app()
